Made some changes, bumped version, added examapample

+++ source/schedule.hxx
@@ -11,8 +11,8 @@ struct _slot_t;
 struct Kernel::Schedule::_sched_t
 {
     _slot_t *slots = nullptr;
-    uint64_t num_slots = 0;
-    uint64_t occupied = 0;
+    uint32_t num_slots = 0;
+    uint32_t occupied = 0;
     uint64_t idctr = 1;
     void run(uint64_t, uint64_t);
 };
